+++ src/main/java/tornadofx/Workspace.kt
@@ -93,6 +93,7 @@ open class Workspace(title: String = "Workspace", navigationMode: NavigationMode
         var defaultSavable = true
         var defaultDeletable = true
         var defaultRefreshable = true
+        var defaultCloseable = true
 
         init {
             FX.log.warning("The Workspace feature is experimental and subject to change even in minor releases!")